% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DragAndDrop.R
\name{DragAndDrop}
\alias{DragAndDrop}
\title{<Add Title>}
\usage{
DragAndDrop(
  ...,
  fileName = NULL,
  textFile = NULL,
  codeFile = NULL,
  sectionFile = NULL,
  width = "100\%",
  height = "100\%",
  elementId = NULL
)
}
\description{
<Add Description>
}
